extract node and link components
